Check allowed hosts before checking the `request_interceptor`
This commit updates the `request_interceptor` implementation of outbound
HTTP requests to check the set of allowed hosts before calling the
`request_interceptor`. This fixes behavior in Spin where service-chained
requests were allowed regardless of configuration of
`allowed_outbound_hosts` in `spin.toml`. A new test is added to ensure
that a request to the "back" component is disallowed.

Signed-off-by: Alex Crichton <alex@alexcrichton.com>