Optimize per-request work dispatching to wasm
This commit implements an optimization available from Wasmtime to
improve the per-request dispatching to a wasm handler for an HTTP
request. Previously the typed view of a request handler was loaded on
each request but this involves somewhat expensive work such as:

* A number of string lookups to find the right export.
* A type-check to ensure that the component export has the right type.

This can all be front-loaded to component-load time with `*Indices`
structures that are generated by Wasmtime's `bindgen!` macro. This
commit uses these generated structures to replace handler detection and
then plumbs the final `*Indices` structure to dispatching the request so
the dispatch doesn't need to perform extra work per-request.

Signed-off-by: Alex Crichton <alex@alexcrichton.com>