Chore(governance): Propose adding js-wasi-ext
This commit adds the js-wasi-ext to the spin project. This is a library
compatible with WASI 0.2 that makes select Node.js APIs available in
ComponentizeJS and StarlingMonkey. This library depends on the build
tooling used in the Spin JS SDK. This enables having more library
compatibility for npm packages with the Spin JS SDK.

Signed-off-by: Karthik Ganeshram <karthik.ganeshram@fermyon.com>