Refactor how `InitContext` works with factors
This commit is a refactoring to replace the concrete `InitContext`
structure with a generic type parameter instead. The motivation for this
is to prepare for handling bytecodealliance/wasmtime#10770. That's a big
change to how `add_to_linker` functions work, notably that the argument
to the generated functions is a `fn`, not a `F: Fn`. WASI factors
currently rely on the closure-like nature this argument which means it's
not compatible with that change. The refactoring to use a `trait
InitContext` here enables plumbing a type parameter through a function
to be able to get a function pointer without relying on closures.

Signed-off-by: Alex Crichton <alex@alexcrichton.com>