Update to Wasmtime 34.0.0
This commit updates to the [34.0.0 release of Wasmtime][rel] where the
main impact in terms of API changes is the introduction of the `HasData`
trait and refactoring of `add_to_linker*` functions. This now requires
type annotations on bindgen-generated `add_to_linker` functions
throughout the codebase. Another minor impact is that Wasmtime now
requires `'static` to be specified on a few locations `T` shows up for
`Store<T>`, but this was just a few small annotations here and there
within Spin.

[rel]: https://github.com/bytecodealliance/wasmtime/releases/tag/v34.0.0

Signed-off-by: Alex Crichton <alex@alexcrichton.com>